only Project Managers can CRUD projects now

+++ src/main/java/org/example/pmas/controller/ProjectController.java
@@ -16,16 +16,20 @@ import java.util.List;
 @RequestMapping("/projects")
 public class ProjectController {
     private final ProjectService projectService;
+    private final SessionHandler sessionHandler;
 
 
-    public ProjectController(ProjectService projectService){
-        this.projectService=projectService;
+    public ProjectController(ProjectService projectService, SessionHandler sessionHandler){
+        this.projectService = projectService;
+        this.sessionHandler = sessionHandler;
     }
 
     //TODO: add session stuff
     @GetMapping("/create-project-form")
     public String getCreateProjectForm(Model model){
         Project project = new Project();
+        boolean allowAccess = isUserProjectManager();
+        model.addAttribute("allowAccess", allowAccess);
         model.addAttribute("project", project);
 
         return "create-project-form";
@@ -34,8 +38,10 @@ public class ProjectController {
     //TODO: add session stuff, redirect to somewhere better
     @PostMapping("/create-project")
     public String createProject(@ModelAttribute("project") Project project){
+        if(isUserProjectManager()){
+            projectService.createProject(project);
+        }
 
-        projectService.createProject(project);
 
         return "redirect:see-all";
     }
@@ -43,8 +49,11 @@ public class ProjectController {
     //TODO: ADD stuff so only the cto can see this page
     @GetMapping("/see-all")
     public String seeAll(Model model){
+
+        boolean allowAccess = isUserProjectManager();
         List<Project> projects = projectService.readAll();
         model.addAttribute("projects", projects);
+        model.addAttribute("allowAccess", allowAccess);
 
         return "show-all-projects";
     }
@@ -52,11 +61,10 @@ public class ProjectController {
     @GetMapping("/my-projects")
     public String myProjects(Model model, HttpSession session){
 
-        var foundObject = session.getAttribute("user");
-        User user = null;
+
+        User user = sessionHandler.getCurrentUser();
         boolean loggedIn = false;
-        if(foundObject instanceof User){
-            user = (User) foundObject;
+        if(user != null){
             loggedIn = true;
         }
         List<Project> projects = null;
@@ -85,7 +93,9 @@ public class ProjectController {
         }else{
 
             Project project = projectService.readSelected(id);
-            model.addAttribute(project);
+            model.addAttribute("project", project);
+            boolean allowAccess = isUserProjectManager();
+            model.addAttribute("allowAccess",allowAccess);
             return "update-project-form";
         }
         //Project project = projectService.
@@ -95,18 +105,18 @@ public class ProjectController {
     public String updateProject(@ModelAttribute Project project, Model model){
 
 
-        //until i figure out how to send both the old project and the new project,
-        //we just send the new project twice. all we need from the old project is
-        //the id, which should be the same as the new project no matter what.
-        projectService.updateProject(project);
-
+        if(isUserProjectManager()){
+            projectService.updateProject(project);
+        }
 
         return "redirect:see-all";
     }
 
     @PostMapping("delete-project")
     public String deleteProject(@ModelAttribute Project project, Model model){
-        projectService.deleteProject(project.getId());
+        if(isUserProjectManager()){
+            projectService.deleteProject(project.getId());
+        }
         return "redirect:see-all";
     }
 
@@ -117,6 +127,16 @@ public class ProjectController {
         model.addAttribute("subprojects", subprojects);
         model.addAttribute("projectId", projectId);
         return "subprojects-all";
+    }
+
+    private boolean isUserProjectManager(){
+        if(sessionHandler.isLoggedIn()){
+            if(sessionHandler.getUserRole().getName().equals("Project Manager")){
+                return true;
+            }
+        }
+
+        return false;
     }
 
 }
